pyproject.toml
[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[project]
name = "raphael-legal"
version = "0.1.0"
description = "Processador de casos jurídicos com Claude API"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40",
    "python-docx>=1.0",
    "docxtpl>=0.18",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlalchemy>=2.0",
    "python-jose[cryptography]>=3.3",
    "bcrypt>=4.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.27",
]

[tool.setuptools.packages.find]
include = ["raphael_legal*", "api*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
